refactor: create game.py
